shared: Update deserialize to return a size_t

+++ src/shared/serialize.c
@@ -269,7 +269,7 @@ end:
 	return ret;
 }
 
-int buxton_deserialize_message(uint8_t *data, BuxtonControlMessage *r_message,
+size_t buxton_deserialize_message(uint8_t *data, BuxtonControlMessage *r_message,
 			       int size, BuxtonData **list)
 {
 	int offset = 0;
